+++ tests/tutorials/tour/test_simple.py
@@ -1,0 +1,18 @@
+import pytest
+
+from wired import ServiceRegistry
+
+
+@pytest.fixture
+def registry():
+    from tutorials.tour.simple.app import setup
+
+    r: ServiceRegistry = setup()
+    return r
+
+
+def test_greet_a_customer(registry):
+    from tutorials.tour.simple.app import greet_a_customer
+
+    actual = greet_a_customer(registry)
+    assert 'Hello !!' == actual